+++ app/(indoors)/indoor/[slug]/page.tsx
