+++ app/(indoors)/indoors/error.tsx
