+++ app/(indoors)/indoors/loading.tsx
